bug 1

+++ PostProject/PostProject/RegisterCustomerWin.xaml.cs
@@ -148,7 +148,7 @@ namespace PostProject
                 }
                 //
                 string password = generated;
-                int balance = 0;
+                float balance = 0;
 
                 command = "insert into Customer values('" + (data.Rows.Count + 1) + "','" + firstName + "','" + lastName + "','" + email + "','" + ssn + "','" + phoneNumber + "','" + username + "','" + password + "','" + balance + "')";
                 SqlCommand cmd = new(command, conn);
